enemy now uses fuel particle explosion and is a PhysicActor
